+++ apps/native-ui-storybook/babel.config.js
@@ -9,6 +9,10 @@ module.exports = function (api) {
         {
           alias: {
             '@utilitywarehouse/native-ui': path.join(__dirname, '../../packages/native-ui/src'),
+            '@utilitywarehouse/native-ui/lab': path.join(
+              __dirname,
+              '../../packages/native-ui/src/lab'
+            ),
             '@utilitywarehouse/native-ui/unstyled': path.join(
               __dirname,
               '../../packages/native-ui/src/components/unstyled'