+++ packages/native-ui/src/config/gluestack-ui.config.ts
@@ -187,7 +187,7 @@ export const gluestackUIConfig = createConfig({
     variants: {
       hardShadow: {
         '1': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: -2,
             height: 2,
@@ -197,7 +197,7 @@ export const gluestackUIConfig = createConfig({
           elevation: 10,
         },
         '2': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 3,
@@ -207,7 +207,7 @@ export const gluestackUIConfig = createConfig({
           elevation: 10,
         },
         '3': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 2,
             height: 2,
@@ -217,7 +217,7 @@ export const gluestackUIConfig = createConfig({
           elevation: 10,
         },
         '4': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: -3,
@@ -226,10 +226,8 @@ export const gluestackUIConfig = createConfig({
           shadowOpacity: 0.5,
           elevation: 10,
         },
-        // this 5th version is only for toast shadow
-        // temporary
         '5': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 3,
@@ -241,7 +239,7 @@ export const gluestackUIConfig = createConfig({
       },
       softShadow: {
         '1': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 0,
@@ -255,7 +253,7 @@ export const gluestackUIConfig = createConfig({
           },
         },
         '2': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 0,
@@ -270,7 +268,7 @@ export const gluestackUIConfig = createConfig({
           },
         },
         '3': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 0,
@@ -285,7 +283,7 @@ export const gluestackUIConfig = createConfig({
           },
         },
         '4': {
-          shadowColor: '$backgroundLight900',
+          shadowColor: '$grey900',
           shadowOffset: {
             width: 0,
             height: 0,