+++ packages/native-ui/src/config/theme/Heading.ts
@@ -1,23 +1,18 @@
 import { createStyle } from '@gluestack-style/react';
 import { H1, H2, H3, H4, H5, H6 } from '@expo/html-elements';
 
-// @ts-ignore
-export const Heading: any = createStyle({
-  color: '$textLight900',
-  letterSpacing: '$sm',
+export const Heading = createStyle({
+  color: '$grey1000',
   fontWeight: '$bold',
   fontFamily: '$heading',
-
-  // Overrides expo-html default styling
   marginVertical: 0,
   _dark: {
-    color: '$textDark50',
+    color: '$darkGrey1000',
   },
   variants: {
     isTruncated: {
       true: {
         props: {
-          // @ts-ignore
           numberOfLines: 1,
           ellipsizeMode: 'tail',
         },
@@ -56,62 +51,53 @@ export const Heading: any = createStyle({
     },
     size: {
       '5xl': {
-        //@ts-ignore
         props: { as: H1 },
         fontSize: '$6xl',
         lineHeight: '$7xl',
       },
       '4xl': {
-        //@ts-ignore
         props: { as: H1 },
         fontSize: '$5xl',
         lineHeight: '$6xl',
       },
 
       '3xl': {
-        //@ts-ignore
         props: { as: H1 },
         fontSize: '$4xl',
         lineHeight: '$5xl',
       },
 
       '2xl': {
-        //@ts-ignore
         props: { as: H2 },
         fontSize: '$3xl',
         lineHeight: '$3xl',
       },
 
-      'xl': {
-        //@ts-ignore
+      xl: {
         props: { as: H3 },
         fontSize: '$2xl',
         lineHeight: '$3xl',
       },
 
-      'lg': {
-        //@ts-ignore
+      lg: {
         props: { as: H4 },
         fontSize: '$xl',
         lineHeight: '$2xl',
       },
 
-      'md': {
-        //@ts-ignore
+      md: {
         props: { as: H5 },
         fontSize: '$lg',
         lineHeight: '$lg',
       },
 
-      'sm': {
-        //@ts-ignore
+      sm: {
         props: { as: H6 },
         fontSize: '$md',
         lineHeight: '$lg',
       },
 
-      'xs': {
-        //@ts-ignore
+      xs: {
         props: { as: H6 },
         fontSize: '$sm',
         lineHeight: '$xs',